added Collision Hands Demo
adds collision to the Open_XR_Tools_Hands by using a CharacterBody3D that follows the corresponding XRController3D.

it is able to collide with the world and with static objects

optional:
*	collision on pickables
*	weight on pickables
*	zero_g movement

currently not supported:
* accounting for switching grab points

this scene contains a two handed example
and the collision hands are made to account for this, currently the two handed is mainly useable for ranged weapon objects.

however under its additional settings, there is a optional mod.
if enabled, changing the rotational offset may produce something along the lines of a melee two handed object.