Fix 'create_snap' desync
If the player drops then picks up a 'SnapZone' with a 'PickableObject' in it the 'PickableObject' will lag behind during player or SnapZone movement due to calling 'create_snap'. Adding the driver as a direct child of the grabber prevents any desync. 'create_lerp' function is similar but has been left out until further testing.

Tested on Godot 4.3.stable