Use placed pickable object's collision shapes
When a pickable object is placed on a snap_path the pickable's CollisionShape3Ds are used to detect grabbing from the rail. This behavior is more intuitive for the class user and prevents desync between a temp snap_zone's automatic circle shape and the pickable. It will be more computationally expensive depending on the pickable's shapes.

- Remove errors from missing expected children by using has_node() before get_node()